+++ packages/babel-helper-create-class-features-plugin/src/features.ts
@@ -1,6 +1,6 @@
 import type { File, types as t } from "@babel/core";
 import type { NodePath } from "@babel/traverse";
-import { hasOwnDecorators } from "./decorators.ts";
+import { hasOwnDecorators } from "./decorators-2018-09.ts";
 
 export const FEATURES = Object.freeze(
   process.env.BABEL_8_BREAKING
