+++ packages/babel-helper-create-class-features-plugin/src/index.ts
@@ -14,7 +14,7 @@ import {
   buildCheckInRHS,
 } from "./fields.ts";
 import type { PropPath } from "./fields.ts";
-import { buildDecoratedClass, hasDecorators } from "./decorators.ts";
+import { buildDecoratedClass, hasDecorators } from "./decorators-2018-09.ts";
 import { injectInitialization, extractComputedKeys } from "./misc.ts";
 import {
   enableFeature,
